pyproject.toml
[project]
name = "youtube-transcript-generator"
version = "0.1.0"
description = "Download YouTube audio and generate transcripts using OpenAI Whisper"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts>=0.2.2",
    "click>=8.3.1",
    "openai>=2.15.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
    "yt-dlp>=2025.12.8",
]

[project.scripts]
yt-transcribe = "youtube_transcript_generator.cli:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
